% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-add_flags_and_order.R
\name{add_flags_and_order}
\alias{add_flags_and_order}
\title{add_flags_and_order}
\usage{
add_flags_and_order(fn_df, Std_df, df_peaks, df_AAstd)
}
\arguments{
\item{fn_df}{final version of samples}

\item{Std_df}{final version of standards}

\item{df_peaks}{raw peak data}

\item{df_AAstd}{raw std data}
}
\value{
returns files with full information
}
\description{
Adds flags to know what has a high std, and orders the files in more user based structure.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
