% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_corrections.R
\name{carbon_corrections}
\alias{carbon_corrections}
\title{FUNCTION_TITLE}
\usage{
carbon_corrections(df_AAstd_int, df_peaks_internal)
}
\arguments{
\item{df_AAstd_int}{PARAM_DESCRIPTION}

\item{df_peaks_internal}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
