% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-data_puller_C.R
\name{data_puller_C}
\alias{data_puller_C}
\title{FUNCTION_TITLE}
\usage{
data_puller_C(file_data)
}
\arguments{
\item{file_data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
