% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-df_AAstd_tagger.R
\name{df_AAstd_tagger}
\alias{df_AAstd_tagger}
\title{FUNCTION_TITLE}
\usage{
df_AAstd_tagger(df, AAStd_name)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{AAStd_name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
