% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-id_peaks_in_smps.R
\name{id_peaks_in_smps}
\alias{id_peaks_in_smps}
\title{FUNCTION_TITLE}
\usage{
id_peaks_in_smps(df_all, df_AAstd)
}
\arguments{
\item{df_all}{PARAM_DESCRIPTION}

\item{df_AAstd}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
