% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-Plt_Lab_AAStds_C.R
\name{Plt_Lab_AAStds_C}
\alias{Plt_Lab_AAStds_C}
\title{FUNCTION_TITLE}
\usage{
Plt_Lab_AAStds_C(file_save_AAstds_C)
}
\arguments{
\item{file_save_AAstds_C}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
