% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_CSIA_full_run.R
\name{process_CSIA_full_run}
\alias{process_CSIA_full_run}
\title{FUNCTION_TITLE}
\usage{
process_CSIA_full_run(drift_corr)
}
\arguments{
\item{drift_corr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
