% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessIsodataOuts-package.R
\docType{package}
\name{ProcessIsodataOuts-package}
\alias{ProcessIsodataOuts}
\alias{ProcessIsodataOuts-package}
\title{ProcessIsodataOuts: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Catherine (Catrina) Nowakowski \email{cnowakowski@uri.edu} (\href{https://orcid.org/0000-0002-0372-9556}{ORCID})

}
\keyword{internal}
