% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-read_tag_correct.R
\name{read_tag_correct}
\alias{read_tag_correct}
\title{FUNCTION_TITLE}
\usage{
read_tag_correct(C_N, file_sorc, file_data, AAStd_name)
}
\arguments{
\item{C_N}{PARAM_DESCRIPTION}

\item{file_sorc}{PARAM_DESCRIPTION}

\item{file_data}{PARAM_DESCRIPTION}

\item{AAStd_name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
