% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tag_correct_og.R
\name{read_tag_correct_og}
\alias{read_tag_correct_og}
\title{FUNCTION_TITLE}
\usage{
read_tag_correct_og(C_N, file_sorc, file_data, AAStd_name)
}
\arguments{
\item{C_N}{PARAM_DESCRIPTION}

\item{file_sorc}{PARAM_DESCRIPTION}

\item{file_data}{PARAM_DESCRIPTION}

\item{AAStd_name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
