% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessIsodataOuts-package.R
\docType{package}
\name{ProcessIsodataOuts-package}
\alias{ProcessIsodataOuts}
\alias{ProcessIsodataOuts-package}
\title{ProcessIsodataOuts: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
